now script names make sense
