+++ src/topological_navigation/scripts/map_editor.py
